Fixed releases create function.
