+++ hypin-app/server.js
@@ -17,7 +17,7 @@ const artistsRouter = require('./routes/artists');
 //artist landing page
 // const releasesRouter = require('./routes/releases');
 
-
+const releasesRouter = require('./routes/releases');
 
 const app = express();
 
@@ -51,6 +51,8 @@ app.use('/', indexRouter);
 app.use('/artists', artistsRouter); 
 //HOW will we track the change of this website from 1 to Many ICEBOX ?
 // app.use('/releases', releasesRouter)
+
+app.use('/releases', releasesRouter);
 
 
 // catch 404 and forward to error handler
