+++ hypin-app/views/index.ejs
@@ -1,3 +1,5 @@
 <%- include('../partials/header') %>
 
+
+
 <%- include('../partials/footer') %>
