+++ hypin-app/views/releases/test.ejs
@@ -1,0 +1,7 @@
+<%- include('../partials/header') %>
+
+<form action="/releases/" method="POST">
+    <button type="submit">TEST</button>
+</form>
+
+<%- include('../partials/footer') %>
